Refactor the GameList Repo into 2 Repositories
