+++ Source/GameView/Repositories/GameDetailsRepository/GameDetailsRepoObserver.swift
@@ -1,0 +1,14 @@
+//
+//  GameDetailsRepoObserver.swift
+//  CommonFiles
+//
+//  Created by Tiewhan Smith on 2020/04/22.
+//
+
+import Foundation
+
+public protocol GameDetailsRepoObserver: AnyObject {
+  
+  func gameDetailsFinishedLoading(withData gameList: [Game])
+  
+}
